move weex framework implementation under platforms
